backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
sqlalchemy
python-multipart
pydantic
python-jose[cryptography]
python-dotenv
passlib[bcrypt]